fix: bookmark export - envrionment object not found crash and replace ShareLink with aictivity
